Organize elements in proper typescript files
