+++ src/elements/Input/index.ts
@@ -1,0 +1,3 @@
+import Input from './Input';
+
+export default Input;
